requirements.txt
python-gnupg
configargparse
